backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
scikit-learn
pandas
numpy
biopython
joblib
